+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to DM-X Video Player will be documented in this file.
 - use 'P' instead of 'S' for Parameters window. 'P' can also close the window if opened.
 - remove default video folder settings, save last one used instead.
 - add time signature and bar / beat display (read from .smt cubase tempo track)
+- Allow to move in video based on arrow key (left/right) 
 
 ### 🧩 Misc